refactor: remove deprecated parse_dicom_prototype module
